style: Factorize the variables
For each network part to configure, there were two variables: a boolean
to enable/disable the configuration and the list itself.

Remove the boolean and replace it with a test on the length of the list.
All the lists are empty by default. If the list has items, then it must
be configured. This will lighten the configuration.

The only downside is that it will no longer be possible to declare items
in a list and then disable the configuration of that list, but that does
not make sense IMHO.

Closes: #8